# Conflicts:
#	src/main/java/com/sparta/newsfeedapp/config/WebSecurityConfig.java
#	src/main/java/com/sparta/newsfeedapp/jwt/JwtUtil.java

+++ src/main/java/com/sparta/newsfeedapp/service/CommentService.java
@@ -1,66 +1,49 @@
 package com.sparta.newsfeedapp.service;
 
-import com.sparta.newsfeedapp.dto.CommentRequestDto;
+import com.sparta.newsfeedapp.dto.commentRequestDto.CommentCreateRequestDto;
+import com.sparta.newsfeedapp.dto.commentRequestDto.CommentUpdateRequestDto;
+import com.sparta.newsfeedapp.dto.commentResponseDto.CommentResponseDto;
 import com.sparta.newsfeedapp.entity.Comment;
 import com.sparta.newsfeedapp.entity.Post;
 import com.sparta.newsfeedapp.entity.User;
 import com.sparta.newsfeedapp.repository.CommentRepository;
 import com.sparta.newsfeedapp.repository.PostRepository;
-import com.sparta.newsfeedapp.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
-import java.util.List;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class CommentService {
-    @Autowired
-    CommentRepository commentRepository;
-    @Autowired
-    UserRepository userRepository;
-    @Autowired
-    PostRepository postRepository;
+
+    private final CommentRepository commentRepository;
+    private final PostRepository postRepository;
 
     public Long findMaxIdOfCommentTable(){
         return Optional.ofNullable(commentRepository.findMaxId()).orElse((long)0);
 
     }
 
-    public Comment createNewCommentColum(CommentRequestDto requestDto,
-                                         Long userId,
-                                         Long newsfeedId){
-        Comment comment = new Comment();
-        comment.setContent(requestDto.getContent());
-        comment.setCountLiked(requestDto.getCountLiked());
-
+    public CommentResponseDto createComment(CommentCreateRequestDto requestDto, Long postId, User user){
         // RequestDto > Entity
-
-        User user = userRepository.findById(userId).orElseThrow(NullPointerException::new);
-        comment.setUser(user);
-
-        Post post = postRepository.findById(newsfeedId).orElseThrow(NullPointerException::new);
-        System.out.println(post.getId());
-        comment.setPost(post);
-
+        Post checkPost = postRepository.findById(postId).orElseThrow(NullPointerException::new);
+        Comment comment = new Comment(requestDto, user, checkPost);
         commentRepository.save(comment);
-        return comment;
+        return new CommentResponseDto(comment);
     }
 
-    public List<Comment> getAllComments() {
-        return commentRepository.findAll();
-    }
+    @Transactional
+    public ResponseEntity<String> updateComment(CommentUpdateRequestDto requestDto, Long commentId, User user) {
 
-    public List<Comment> getCommentsBypostId(long newsfeedId){
-        return commentRepository.findByPostId(newsfeedId);
-    }
-
-    public ResponseEntity<String> updateComment(CommentRequestDto requestDto ,Long commentId) {
         if (commentRepository.existsById(commentId)) {
             Comment comment = commentRepository.findById(commentId).orElseThrow(NullPointerException::new);
+            if (!comment.getUser().getId().equals(user.getId())){
+                throw new IllegalArgumentException("본인 댓글만 수정할 수 있습니다.");
+            }
             comment.update(requestDto);
             return new ResponseEntity<>("성공적으로 수정했습니다. (" + comment.getContent() + ")", HttpStatus.OK);
         } else {
@@ -68,18 +51,17 @@ public class CommentService {
         }
     }
 
-    public ResponseEntity<String> deleteComment(Long commentId){
+    public ResponseEntity<String> deleteComment(Long commentId, User user){
 
         if (commentRepository.existsById(commentId)) {
             Comment comment = commentRepository.findById(commentId).orElseThrow(NullPointerException::new);
+            if (!comment.getUser().getId().equals(user.getId())){
+                throw new IllegalArgumentException("본인 댓글만 삭제할 수 있습니다.");
+            }
             commentRepository.deleteById(commentId);
             return new ResponseEntity<>("성공적으로 삭제했습니다.", HttpStatus.OK);
         } else {
             return new ResponseEntity<>("Comment를 찾지 못해 삭제하지 못했습니다.", HttpStatus.NOT_FOUND);
         }
-    }
-
-    public LocalDateTime timeNow(){
-        return LocalDateTime.now();
     }
 }
